add: register feature

+++ app/Http/Controllers/Dashboard/HistoryController.php
@@ -13,12 +13,16 @@ class HistoryController extends Controller
     {
         return view('dashboard.histories.index', [
             'title' => 'History',
-            'histories' => History::latest()->get(),
+            'histories' => History::where('user_id', auth()->user()->id)->latest()->get(),
         ]);
     }
 
     public function show(History $history)
     {
+        if ($history->user_id !== auth()->user()->id) {
+            abort(404);
+        }
+
         return view('dashboard.histories.show', [
             'title' => 'History',
             'history' => $history,
@@ -35,6 +39,10 @@ class HistoryController extends Controller
 
     public function edit(History $history)
     {
+        if ($history->user_id !== auth()->user()->id) {
+            abort(404);
+        }
+
         return view('dashboard.histories.edit', [
             'title' => 'History',
             'history' => $history,
@@ -43,6 +51,10 @@ class HistoryController extends Controller
 
     public function update(UpdateHistoryRequest $request, History $history)
     {
+        if ($history->user_id !== auth()->user()->id) {
+            abort(404);
+        }
+
         $data = $request->validated();
         $history->update($data);
         return redirect($request->previous_url ?? route('dashboard.histories.index'))->with('success', 'Text has been updated');
@@ -50,6 +62,10 @@ class HistoryController extends Controller
 
     public function destroy(History $history)
     {
+        if ($history->user_id !== auth()->user()->id) {
+            abort(404);
+        }
+
         $history->delete();
         return redirect()->back()->with('success', 'Text has been deleted from history');
     }
